Renamed internal/context to internal/reqcontext to avoid name clash with stdlib context
