+++ esignet-service/internal/httpmiddleware/correlationid_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	appcontext "github.com/mosip/esignet/internal/context"
+	"github.com/mosip/esignet/internal/reqcontext"
 )
 
 var uuidV4Pattern = regexp.MustCompile(`^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$`)
@@ -23,7 +23,7 @@ var uuidV4Pattern = regexp.MustCompile(`^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[8
 func TestCorrelationID_GeneratesWhenAbsent(t *testing.T) {
 	var gotFromContext, gotFromHeader string
 	next := http.HandlerFunc(func(_ http.ResponseWriter, r *http.Request) {
-		gotFromContext = appcontext.TraceIDFromContext(r.Context())
+		gotFromContext = reqcontext.TraceIDFromContext(r.Context())
 		gotFromHeader = r.Header.Get(CorrelationIDHeader)
 	})
 
@@ -39,7 +39,7 @@ func TestCorrelationID_GeneratesWhenAbsent(t *testing.T) {
 func TestCorrelationID_ReusesExistingCorrelationIDHeader(t *testing.T) {
 	var got string
 	next := http.HandlerFunc(func(_ http.ResponseWriter, r *http.Request) {
-		got = appcontext.TraceIDFromContext(r.Context())
+		got = reqcontext.TraceIDFromContext(r.Context())
 	})
 
 	req := httptest.NewRequest(http.MethodGet, "/", nil)
@@ -67,7 +67,7 @@ func TestCorrelationID_FallsBackToOtherHeadersInPriorityOrder(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			var got string
 			next := http.HandlerFunc(func(_ http.ResponseWriter, r *http.Request) {
-				got = appcontext.TraceIDFromContext(r.Context())
+				got = reqcontext.TraceIDFromContext(r.Context())
 			})
 
 			req := httptest.NewRequest(http.MethodGet, "/", nil)
@@ -98,7 +98,7 @@ func TestCorrelationID_FallsBackToB3Headers(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			var got string
 			next := http.HandlerFunc(func(_ http.ResponseWriter, r *http.Request) {
-				got = appcontext.TraceIDFromContext(r.Context())
+				got = reqcontext.TraceIDFromContext(r.Context())
 			})
 
 			req := httptest.NewRequest(http.MethodGet, "/", nil)
