+++ esignet-service/internal/reqcontext/reqcontext_test.go
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package context
+package reqcontext
 
 import (
 	"context"
